Data is ready for coding
1. CCES_Data_Cleaning_Final.ipynb:
This is the jupyter notebook I used to import the data, clean it, and write it to a sqlite file. The notebook has good notes throughout to explain all of the processes.

2. data_raw folder:
tab file with survey responses - this is the primary source of raw data.
csv file with state names and a common column V101 with the tab file.

3. db\CCES_Ver50.sqlite
This is the sqlite file that should be used for all queries.
There are example queries in the jupyter notebook to help with the class, column names, session.query syntax, etc.